refactor: consolidate test suite, remove redundant tests
- Create tests/conftest.py with shared MockBPFState and make_decision
- Remove duplicate MockBPFState / _make_decision from handler test files
- Remove hand-written matching unit tests (TestCidrMatching,
  TestUrlPathMatching, TestHostnameMatching) from test_policy.py — these
  are fully covered by hypothesis property tests in test_policy_properties.py
- Rename test_policy_enforcer.py to test_dns_cache.py (matches actual content)

Net -157 lines, 779 tests still pass.

Closes #127
